feat: SMILESEncoderDataset and SMILESDataset.
